requirements.txt
streamlit
python-dotenv
langchain
langchain-openai
langchain-community
langchain-chroma
chromadb
pypdf
tiktoken
pysqlite3-binary
google-auth-oauthlib
google-auth-httplib2
google-api-python-client